+++ validator/accounts/v2/cmd_accounts.go
@@ -36,6 +36,7 @@ this command outputs a deposit data string which is required to become a validat
 			Flags: []cli.Flag{
 				flags.WalletDirFlag,
 				flags.WalletPasswordsDirFlag,
+				flags.PasswordFileFlag,
 				flags.ShowDepositDataFlag,
 			},
 			Action: func(cliCtx *cli.Context) error {
@@ -51,7 +52,7 @@ this command outputs a deposit data string which is required to become a validat
 			Flags: []cli.Flag{
 				flags.WalletDirFlag,
 				flags.WalletPasswordsDirFlag,
-				flags.BackupPathFlag,
+				flags.BackupDirFlag,
 				flags.AccountsFlag,
 			},
 			Action: func(cliCtx *cli.Context) error {
@@ -67,7 +68,7 @@ this command outputs a deposit data string which is required to become a validat
 			Flags: []cli.Flag{
 				flags.WalletDirFlag,
 				flags.WalletPasswordsDirFlag,
-				flags.BackupPathFlag,
+				flags.BackupDirFlag,
 				flags.PasswordFileFlag,
 			},
 			Action: func(cliCtx *cli.Context) error {
