+++ validator/accounts/v2/import.go
@@ -6,12 +6,10 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"path"
 	"path/filepath"
 	"strings"
 
 	"github.com/logrusorgru/aurora"
-	"github.com/manifoldco/promptui"
 	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/prysm/validator/flags"
 	"github.com/prysmaticlabs/prysm/validator/keymanager/v2/direct"
@@ -34,7 +32,7 @@ func ImportAccount(cliCtx *cli.Context) error {
 		return errors.New("can only export accounts for a non-HD wallet")
 	}
 
-	backupDir, err := inputImportDir(cliCtx)
+	backupDir, err := inputDirectory(cliCtx, importDirPromptText, flags.BackupDirFlag)
 	if err != nil {
 		return errors.Wrap(err, "could not parse output directory")
 	}
@@ -61,26 +59,6 @@ func ImportAccount(cliCtx *cli.Context) error {
 	}
 
 	return nil
-}
-
-func inputImportDir(cliCtx *cli.Context) (string, error) {
-	outputDir := cliCtx.String(flags.BackupPathFlag.Name)
-	if cliCtx.IsSet(flags.BackupPathFlag.Name) {
-		return outputDir, nil
-	}
-	if outputDir == flags.DefaultValidatorDir() {
-		outputDir = path.Join(outputDir)
-	}
-	prompt := promptui.Prompt{
-		Label:    "Enter the file location of the exported wallet zip to import",
-		Validate: validateDirectoryPath,
-		Default:  outputDir,
-	}
-	outputPath, err := prompt.Run()
-	if err != nil {
-		return "", fmt.Errorf("could not determine import directory: %v", formatPromptError(err))
-	}
-	return outputPath, nil
 }
 
 func unzipArchiveToTarget(archiveDir string, target string) ([]string, error) {
