+++ validator/accounts/v2/wallet_create.go
@@ -2,10 +2,7 @@ package v2
 
 import (
 	"context"
-	"fmt"
-	"strings"
 
-	"github.com/manifoldco/promptui"
 	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/prysm/validator/flags"
 	v2keymanager "github.com/prysmaticlabs/prysm/validator/keymanager/v2"
@@ -30,7 +27,7 @@ func CreateWallet(cliCtx *cli.Context) error {
 		}
 		log.WithField("wallet-path", w.accountsPath).Infof(
 			"Successfully created wallet with on-disk keymanager configuration. " +
-				"Make a new validator account with ./prysm.sh validator accounts-2 new",
+				"Make a new validator account with ./prysm.sh validator accounts-v2 create",
 		)
 	case v2keymanager.Derived:
 		if err = createDerivedKeymanagerWallet(cliCtx, w); err != nil {
@@ -38,7 +35,7 @@ func CreateWallet(cliCtx *cli.Context) error {
 		}
 		log.WithField("wallet-path", w.accountsPath).Infof(
 			"Successfully created HD wallet and saved configuration to disk. " +
-				"Make a new validator account with ./prysm.sh validator accounts-2 new",
+				"Make a new validator account with ./prysm.sh validator accounts-2 create",
 		)
 	case v2keymanager.Remote:
 		if err = createRemoteKeymanagerWallet(cliCtx, w); err != nil {
@@ -67,7 +64,7 @@ func createDirectKeymanagerWallet(cliCtx *cli.Context, wallet *Wallet) error {
 func createDerivedKeymanagerWallet(cliCtx *cli.Context, wallet *Wallet) error {
 	skipMnemonicConfirm := cliCtx.Bool(flags.SkipMnemonicConfirmFlag.Name)
 	ctx := context.Background()
-	walletPassword, err := inputNewWalletPassword(cliCtx)
+	walletPassword, err := inputPassword(cliCtx, newWalletPasswordPromptText, confirmPass)
 	if err != nil {
 		return err
 	}
@@ -104,84 +101,6 @@ func createRemoteKeymanagerWallet(cliCtx *cli.Context, wallet *Wallet) error {
 	}
 	if err := wallet.WriteKeymanagerConfigToDisk(ctx, keymanagerConfig); err != nil {
 		return errors.Wrap(err, "could not write keymanager config to disk")
-	}
-	return nil
-}
-
-func inputRemoteKeymanagerConfig(cliCtx *cli.Context) (*remote.Config, error) {
-	addr := cliCtx.String(flags.GrpcRemoteAddressFlag.Name)
-	crt := cliCtx.String(flags.RemoteSignerCertPathFlag.Name)
-	key := cliCtx.String(flags.RemoteSignerKeyPathFlag.Name)
-	ca := cliCtx.String(flags.RemoteSignerCACertPathFlag.Name)
-	if addr != "" && crt != "" && key != "" && ca != "" {
-		newCfg := &remote.Config{
-			RemoteCertificate: &remote.CertificateConfig{
-				ClientCertPath: strings.TrimRight(crt, "\r\n"),
-				ClientKeyPath:  strings.TrimRight(key, "\r\n"),
-				CACertPath:     strings.TrimRight(ca, "\r\n"),
-			},
-			RemoteAddr: strings.TrimRight(addr, "\r\n"),
-		}
-		log.Infof("New configuration")
-		fmt.Printf("%s\n", newCfg)
-		return newCfg, nil
-	}
-	log.Infof("Input desired configuration")
-	prompt := promptui.Prompt{
-		Label: "Remote gRPC address (such as host.example.com:4000)",
-		Validate: func(input string) error {
-			if input == "" {
-				return errors.New("remote host address cannot be empty")
-			}
-			return nil
-		},
-	}
-	remoteAddr, err := prompt.Run()
-	if err != nil {
-		return nil, err
-	}
-	prompt = promptui.Prompt{
-		Label:    "Path to TLS crt (such as /path/to/client.crt)",
-		Validate: validateCertPath,
-	}
-	clientCrtPath, err := prompt.Run()
-	if err != nil {
-		return nil, err
-	}
-	prompt = promptui.Prompt{
-		Label:    "Path to TLS key (such as /path/to/client.key)",
-		Validate: validateCertPath,
-	}
-	clientKeyPath, err := prompt.Run()
-	if err != nil {
-		return nil, err
-	}
-	prompt = promptui.Prompt{
-		Label:    "(Optional) Path to certificate authority (CA) crt (such as /path/to/ca.crt)",
-		Validate: validateCertPath,
-	}
-	caCrtPath, err := prompt.Run()
-	if err != nil {
-		return nil, err
-	}
-	newCfg := &remote.Config{
-		RemoteCertificate: &remote.CertificateConfig{
-			ClientCertPath: strings.TrimRight(clientCrtPath, "\r\n"),
-			ClientKeyPath:  strings.TrimRight(clientKeyPath, "\r\n"),
-			CACertPath:     strings.TrimRight(caCrtPath, "\r\n"),
-		},
-		RemoteAddr: strings.TrimRight(remoteAddr, "\r\n"),
-	}
-	fmt.Printf("%s\n", newCfg)
-	return newCfg, nil
-}
-
-func validateCertPath(input string) error {
-	if input == "" {
-		return errors.New("crt path cannot be empty")
-	}
-	if !fileExists(input) {
-		return fmt.Errorf("no crt found at path: %s", input)
 	}
 	return nil
 }
