+++ validator/accounts/v2/wallet_test.go
@@ -7,7 +7,7 @@ import (
 	"io/ioutil"
 	"math/big"
 	"os"
-	"path"
+	"path/filepath"
 	"testing"
 
 	"github.com/prysmaticlabs/prysm/shared/testutil"
@@ -42,13 +42,13 @@ func setupWalletCtx(
 	set.String(flags.WalletDirFlag.Name, cfg.walletDir, "")
 	set.String(flags.WalletPasswordsDirFlag.Name, cfg.passwordsDir, "")
 	set.String(flags.KeymanagerKindFlag.Name, cfg.keymanagerKind.String(), "")
-	set.String(flags.BackupPathFlag.Name, cfg.exportDir, "")
+	set.String(flags.BackupDirFlag.Name, cfg.exportDir, "")
 	set.String(flags.AccountsFlag.Name, cfg.accountsToExport, "")
 	set.String(flags.PasswordFileFlag.Name, cfg.passwordFile, "")
 	assert.NoError(tb, set.Set(flags.WalletDirFlag.Name, cfg.walletDir))
 	assert.NoError(tb, set.Set(flags.WalletPasswordsDirFlag.Name, cfg.passwordsDir))
 	assert.NoError(tb, set.Set(flags.KeymanagerKindFlag.Name, cfg.keymanagerKind.String()))
-	assert.NoError(tb, set.Set(flags.BackupPathFlag.Name, cfg.exportDir))
+	assert.NoError(tb, set.Set(flags.BackupDirFlag.Name, cfg.exportDir))
 	assert.NoError(tb, set.Set(flags.AccountsFlag.Name, cfg.accountsToExport))
 	assert.NoError(tb, set.Set(flags.PasswordFileFlag.Name, cfg.passwordFile))
 	return cli.NewContext(&app, set, nil)
@@ -57,9 +57,9 @@ func setupWalletCtx(
 func setupWalletAndPasswordsDir(t testing.TB) (string, string) {
 	randPath, err := rand.Int(rand.Reader, big.NewInt(1000000))
 	require.NoError(t, err, "Could not generate random file path")
-	walletDir := path.Join(testutil.TempDir(), fmt.Sprintf("/%d", randPath), "wallet")
+	walletDir := filepath.Join(testutil.TempDir(), fmt.Sprintf("/%d", randPath), "wallet")
 	require.NoError(t, os.RemoveAll(walletDir), "Failed to remove directory")
-	passwordsDir := path.Join(testutil.TempDir(), fmt.Sprintf("/%d", randPath), "passwords")
+	passwordsDir := filepath.Join(testutil.TempDir(), fmt.Sprintf("/%d", randPath), "passwords")
 	require.NoError(t, os.RemoveAll(passwordsDir), "Failed to remove directory")
 	t.Cleanup(func() {
 		require.NoError(t, os.RemoveAll(walletDir), "Failed to remove directory")
